Extract UI data from Draft to NewMessageVM
